chore(performance): move from 64 to 32 floats

+++ causy/graph.py
@@ -304,7 +304,7 @@ class Graph(BaseGraphInterface):
             raise ValueError(f"Node with id {id_} already exists")
 
         try:
-            tensor_values = torch.tensor(values, dtype=torch.float64)
+            tensor_values = torch.tensor(values, dtype=torch.float32)
         except TypeError as e:
             raise ValueError(f"Currently only numeric values are supported. {e}")
